Move UIImage with Color to extension
